fix memory leak on yaml set input string
previously we don't call kk_string_drop, because if we do it will make
the yaml parser parse on freed string. I deliberately leaks the string
just to see the parser work.

This PR fixes it, the yaml parser now borrow koka string. And guaranteed
the string valid until the parser freed.